0722 エスケープシーケンスとフラグ学習

+++ 第15回_実行時引数/配布用/AverageCalculator.java
@@ -2,10 +2,10 @@
 // 実行例：java AverageCalculator.java 10 20 30 （→ Average: 20.0）
 class AverageCalculator {
     public static void main(String[] args) {
-        int num1 = Integer.parseInt(args[0]);
+        double num1 = Double.parseDouble(args[0]);
         int num2 = Integer.parseInt(args[1]);
         int num3 = Integer.parseInt(args[2]);
-        double average = (num1 + num2 + num3) / 3.0;
+        double average = (num1 + num2 + num3) / 3;
         System.out.print("Average: " + average);
     }
 }
